almost done with settings page just need to fix caching issue

+++ server/controllers/users.js
@@ -1,17 +1,26 @@
-import { User } from "../models/User";
+import { User } from "../models/User.js";
+import bcrypt from "bcrypt";
+import fs from "fs";
 
 export const patchSettings = async (req, res) => {
     try{
         const {userId} = req.params;
-        const updates = req.body;
+        const fields = ['firstName', 'lastName', 'username', 'email', 'password'];
+        const updates = Object.fromEntries(Object.entries(req.body)
+                            .filter(([key, val]) => fields.includes(key) && val));
 
-        //make sure if client sent null, then multer didn't upload
-        //and if client sent a picture path, then multer did upload and with the same path
-        if(updates.picturePath != req.dbPicturePath){
-            throw new Error("Invalid picture path");
+
+        
+        if(req.dbPicturePath){
+            updates.picturePath = req.dbPicturePath
+        }
+        if(updates.password){
+            const salt = await bcrypt.genSalt();
+            const passwordHash = await bcrypt.hash(updates.password, salt);
+            updates.password = passwordHash;
         }
         const updatedUser = await User.findByIdAndUpdate(userId, updates, {new: true});
-        if(updates.picturePath){
+        if(req.curPicturePath){
             fs.rename(req.curPicturePath, req.dbPicturePath, (err) => {
                 //this shouldn't ever error, but if it did it would just result in the user having no pfp
                 //which they could change in settings
@@ -26,12 +35,14 @@ export const patchSettings = async (req, res) => {
                 console.log(`Renamed ${req.curPicturePath} to ${req.dbPicturePath}`);
             });
         }
-        return res.status(200).json(updatedUser);
+        return res.status(200).json({user: updatedUser});
     } catch(err){
-        fs.unlink(req.curPicturePath, (err) => {
-            if(err) return console.log(err);
-            console.log(`${req.curPicturePath} was deleted`);
-        });
+        if(req.curPicturePath){
+            fs.unlink(req.curPicturePath, (err) => {
+                if(err) return console.log(err);
+                console.log(`${req.curPicturePath} was deleted`);
+            });
+        }
         return res.status(409).json({error: err.message});
     }
